Factor out a common TestHtmlParser
Looks like I'm doing this in a few different places.
